python_services/requirements.txt
pillow==10.4.0
openai==1.52.2
opencv-python==4.10.0.84
numpy==1.26.4
requests==2.31.0
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.12
aiofiles==23.2.1
python-dotenv==1.0.1
scikit-image==0.24.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
torch==2.5.1
torchvision==0.20.1
transformers==4.46.2
diffusers==0.31.0
accelerate==1.1.1